requirements.txt
alembic==1.9.4
click==8.1.3
colorama==0.4.6
Flask~=2.2.2
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.2
SQLAlchemy~=2.0.3
typing_extensions==4.5.0
Werkzeug~=2.2.2

pip~=23.0
attrs~=22.2.0
Pillow~=9.4.0
visitor~=0.1.3
python-dateutil~=2.8.2
setuptools~=65.5.0
packaging~=23.0
pyparsing~=3.0.9
mypy~=1.0.0
WTForms~=3.0.1
python-dotenv~=1.0.0